- event.c:event_run(): changed 'len' variable to an unsigned int.   Note that dlink_list_length() is unsigned int.
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@4119 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/event.c
@@ -76,7 +76,7 @@ void
 event_run(void)
 {
   static time_t last = 0;
-  size_t len = 0;
+  unsigned int len = 0;
 
   if (last == CurrentTime)
     return;
